requirements.txt
gradio
pytesseract
pillow
